reorganization part 2
