+++ openmsipython/data_models/laser_shock/epoxy_ID.py
@@ -1,8 +1,8 @@
 #imports
 from gemd.entity.value import NominalCategorical
+from ..spec_from_filemaker_record import MaterialSpecFromFileMakerRecord
 from .attribute_templates import ATTR_TEMPL
 from .object_templates import OBJ_TEMPL
-from .spec_from_filemaker_record import MaterialSpecFromFileMakerRecord
 
 class LaserShockEpoxyID(MaterialSpecFromFileMakerRecord) :
     """
